backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.1.2",
    "django>=6.0.3",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.17.1",
    "openai-agents>=0.13.4",
    "pinecone>=8.1.1",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.2",
    "reportlab>=4.4.10",
    "requests>=2.33.1",
]
